correct alloc_handle prototype

+++ libctru/source/services/soc/soc_common.h
@@ -10,7 +10,7 @@
 
 #define SYNC_ERROR ENODEV
 
-int __alloc_handle(int size);
+int __alloc_handle(size_t size);
 __handle *__get_handle(int fd);
 void __release_handle(int fd);
